refactor(database): align module layout with standards
